pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.13.0",
    "ruff>=0.12.11",
]
